Refactored linters into separate package/class.
- Linters have a common superclass `Linter` that provides machinery for converting from LinterResult
into Check.
- Linter WARNINGS are now ignored, as they don't map to a CHECK result. This will probably change in
the future.
- Moved linter implementation to separate file.
- Renamed result.py to lint.py for linting base classes (Linter, LintSeverity, &c)
