+++ datacontract/lint/resolve.py
@@ -10,8 +10,7 @@ from datacontract.lint.urls import fetch_resource
 from datacontract.model.data_contract_specification import \
     DataContractSpecification
 from datacontract.model.exceptions import DataContractException
-from .result import LinterResult
-from .linters import lint_all
+
 
 def resolve_data_contract(
     data_contract_location: str = None,
@@ -32,9 +31,6 @@ def resolve_data_contract(
             reason=f"Data contract needs to be provided",
             engine="datacontract",
         )
-
-def lint_data_contract(data_contract_yaml) -> LinterResult:
-    lint_all(data_contract_yaml)
 
 
 def resolve_data_contract_from_location(location) -> DataContractSpecification:
