+++ tests/test_linters.py
@@ -1,26 +1,41 @@
-import datacontract.lint.linters as linters
+from datacontract.lint.linters.example_model_linter import ExampleModelLinter
 import datacontract.lint.resolve as resolve
-import datacontract.lint.result as lint_result
+import datacontract.lint.lint as lint
 import datacontract.model.data_contract_specification as spec
+from datacontract.model.run import Check
+
+
+def construct_error_check(msg: str) -> Check:
+    return Check(
+        type="lint",
+        name="Linter 'Example(s) match model'",
+        result="warning",
+        engine="datacontract",
+        reason=msg
+    )
+
+
+success_check = Check(
+    type="lint",
+    name="Linter 'Example(s) match model'",
+    result="passed",
+    engine="datacontract"
+)
 
 
 def test_lint_invalid_model():
     data_contract_file = "examples/lint/datacontract_unknown_model.yaml"
     contract = resolve.resolve_data_contract_from_location(data_contract_file)
-    result = linters.linter_examples_conform_to_model(contract)
-    expected = lint_result.LinterMessage.error(
-        "Example 1 has non-existent model 'orders'")
-    assert result.error_results() == [
-        expected
-    ]
-    assert result.warning_results() == []
+    result = ExampleModelLinter().lint(contract)
+    expected = construct_error_check("Example 1 has non-existent model 'orders'")
+    assert result == [expected]
 
 
 def test_lint_valid_csv_columns():
     contract = resolve.resolve_data_contract_from_location(
         "examples/lint/datacontract_csv_lint_base.yaml")
-    result = linters.linter_examples_conform_to_model(contract)
-    assert result.no_errors_or_warnings()
+    result = ExampleModelLinter().lint(contract)
+    assert result == [success_check]
 
 
 def test_lint_extra_model_columns():
@@ -31,13 +46,12 @@ def test_lint_extra_model_columns():
         "column_2": spec.Field(type="str"),
         "column_3": spec.Field(type="str", required=True)
     })
-    result = linters.linter_examples_conform_to_model(base_spec)
-    expected = lint_result.LinterMessage.error(
+    result = ExampleModelLinter().lint(base_spec)
+    expected = construct_error_check(
         "Example 1 is missing field 'column_3' required by model 'orders'")
-    assert result.error_results() == [
+    assert result == [
         expected
     ]
-    assert result.warning_results() == []
 
 
 def test_lint_extra_example_columns():
@@ -47,13 +61,12 @@ def test_lint_extra_example_columns():
         type="csv",
         model="orders",
         data="column_1, column_2, column_3\nvalue_1, value_2, value_3")
-    result = linters.linter_examples_conform_to_model(base_spec)
-    expected = lint_result.LinterMessage.error(
+    result = ExampleModelLinter().lint(base_spec)
+    expected = construct_error_check(
         "Example 1 has field 'column_3' that's not contained in model 'orders'")
-    assert result.error_results() == [
+    assert result == [
         expected
     ]
-    assert result.warning_results() == []
 
 
 def test_lint_json_example():
@@ -63,8 +76,8 @@ def test_lint_json_example():
         type="json",
         model="orders",
         data='{"column_1": 1, "column_2": 2}')
-    result = linters.linter_examples_conform_to_model(base_spec)
-    assert result.no_errors_or_warnings()
+    result = ExampleModelLinter().lint(base_spec)
+    assert result == [success_check]
 
 
 def test_lint_json_example_extra_columns():
@@ -74,13 +87,11 @@ def test_lint_json_example_extra_columns():
         type="json",
         model="orders",
         data='{"column_1": 1, "column_2": 2, "column_3": 3}')
-    result = linters.linter_examples_conform_to_model(base_spec)
-    expected = lint_result.LinterMessage.error(
+    result = ExampleModelLinter().lint(base_spec)
+    expected = construct_error_check(
         "Example 1 has field 'column_3' that's not contained in model 'orders'"
-        )
-    assert result.error_results() == [
-        expected
-    ]
+    )
+    assert result == [expected]
 
 
 def test_lint_yaml_example():
@@ -93,5 +104,5 @@ def test_lint_yaml_example():
 column_1: 1
 column_2: 2
 """)
-    result = linters.linter_examples_conform_to_model(base_spec)
-    assert result.no_errors_or_warnings()
+    result = ExampleModelLinter().lint(base_spec)
+    assert result == [success_check]
